requirements.txt
pytest
pytest-asyncio
pillow
python-dotenv
twikit
twikit_ext
proxystr
ruff
